Restructure build.js into focused modules
Split the 2,585-line monolithic build.js into separate modules for
branding, icons, helpers, renderers, and per-part slide data. The
orchestrator build.js is now ~30 lines. Output is identical (49 slides,
all content preserved).
